Move ProfileReader to profile module
