+++ generator/src/main/java/com/scottlogic/deg/generator/commandline/GenerateCommandLine.java
@@ -11,7 +11,7 @@ import java.nio.file.Path;
 import java.util.List;
 
 import static com.scottlogic.deg.generator.config.detail.CombinationStrategyType.MINIMAL;
-import static com.scottlogic.deg.generator.generation.Defaults.DEFAULT_MAX_ROWS;
+import static com.scottlogic.deg.common.util.Defaults.DEFAULT_MAX_ROWS;
 import static com.scottlogic.deg.generator.config.detail.DataGenerationType.RANDOM;
 import static com.scottlogic.deg.generator.config.detail.OutputFormat.CSV;
 import static com.scottlogic.deg.generator.config.detail.TreeWalkerType.REDUCTIVE;
