+++ generator/src/main/java/com/scottlogic/deg/generator/outputs/formats/csv/CsvOutputFormat.java
@@ -1,12 +1,12 @@
 package com.scottlogic.deg.generator.outputs.formats.csv;
 
 import com.scottlogic.deg.common.profile.ProfileFields;
-import com.scottlogic.deg.generator.outputs.formats.DataSetWriter;
-import com.scottlogic.deg.generator.outputs.formats.OutputFormat;
+    import com.scottlogic.deg.generator.outputs.formats.DataSetWriter;
+    import com.scottlogic.deg.generator.outputs.formats.OutputFormat;
 
-import java.io.IOException;
-import java.io.OutputStream;
-import java.util.Optional;
+    import java.io.IOException;
+    import java.io.OutputStream;
+    import java.util.Optional;
 
 public class CsvOutputFormat implements OutputFormat {
     @Override
