+++ generator/src/test/java/com/scottlogic/deg/generator/cucumber/testframework/steps/GeneralTestStep.java
@@ -5,7 +5,7 @@ import com.scottlogic.deg.generator.config.detail.CombinationStrategyType;
 import com.scottlogic.deg.generator.config.detail.DataGenerationType;
 import com.scottlogic.deg.generator.config.detail.TreeWalkerType;
 import com.scottlogic.deg.generator.cucumber.testframework.utils.*;
-import com.scottlogic.deg.generator.inputs.InvalidProfileException;
+import com.scottlogic.deg.profile.reader.InvalidProfileException;
 import com.scottlogic.deg.profile.v0_1.AtomicConstraintType;
 import cucumber.api.java.Before;
 import cucumber.api.java.en.*;
