+++ generator/src/test/java/com/scottlogic/deg/generator/cucumber/testframework/steps/JSONTestStep.java
@@ -1,7 +1,7 @@
 package com.scottlogic.deg.generator.cucumber.testframework.steps;
 
 import com.scottlogic.deg.generator.cucumber.testframework.utils.CucumberTestState;
-import com.scottlogic.deg.generator.inputs.InvalidProfileException;
+import com.scottlogic.deg.profile.reader.InvalidProfileException;
 import cucumber.api.java.en.And;
 
 import java.io.IOException;
