+++ generator/src/test/java/com/scottlogic/deg/generator/cucumber/testframework/utils/CucumberProfileReader.java
@@ -6,9 +6,9 @@ import com.scottlogic.deg.common.profile.Profile;
 import com.scottlogic.deg.common.profile.ProfileFields;
 import com.scottlogic.deg.common.profile.Rule;
 import com.scottlogic.deg.common.profile.constraints.Constraint;
-import com.scottlogic.deg.generator.inputs.InvalidProfileException;
-import com.scottlogic.deg.generator.inputs.MainConstraintReader;
-import com.scottlogic.deg.generator.inputs.ProfileReader;
+import com.scottlogic.deg.profile.reader.InvalidProfileException;
+import com.scottlogic.deg.profile.reader.MainConstraintReader;
+import com.scottlogic.deg.profile.reader.ProfileReader;
 import com.scottlogic.deg.common.profile.RuleInformation;
 
 import java.nio.file.Path;
