+++ profile/build.gradle
@@ -3,6 +3,7 @@
  */
 
 dependencies {
+    compile project(':common')
     compile 'org.jetbrains.kotlin:kotlin-stdlib:1.3.21'
     compile 'com.worldturner.medeia:medeia-validator-jackson:1.0.0'
     compile 'org.leadpony.justify:justify:0.14.0'
