+++ profile/src/main/java/com/scottlogic/deg/profile/reader/InvalidProfileException.java
@@ -1,4 +1,4 @@
-package com.scottlogic.deg.generator.inputs;
+package com.scottlogic.deg.profile.reader;
 
 public class InvalidProfileException extends Exception
 {
